requirements.txt
mindsdb-sdk
python-dotenv
langchain
langchain-community
langchain-weaviate
langchain-text-splitters
langchain-google-genai
weaviate-client
unstructured
unstructured[md]
unstructured[rst]
streamlit

# pandoc